ENH: Expanded test coverage for LS models
Added test coverage for some corner cases of WLS, which revealed a bug
in WLS.
Added test coverage for norm_resid, which was also incorrect.
Added exception checks for GLS sigma with wrong sizes.
Added test for QR where Q,R are user provided.  Revealed a bug in this code.