ENH: Alternative travis script that uses Anaconda via Miniconda
Provides an alternative method to test on Travis using Miniconda
that has some advantages from the current system.

- All binary, so no time spent building
- No branching in the execution steps
- Support for up-to-date requirements which are important to test

Also includes a small change to tools/matplotlibrc which changes the backend
to Agg to avoid Tk-related errors on Travis.  Agg is alays available and does not
depend on Qt or another toolkit.

Closes #1500